+++ packages/client/src/GlobalStyle.tsx
@@ -2,46 +2,44 @@ import { Global, css } from '@emotion/react';
 
 export const GlobalStyle = () => {
   return (
-    <>
-      <Global
-        styles={css`
-          ${reset}
-          * {
-            box-sizing: border-box;
-          }
-          input {
-            outline: none;
-            border: none;
-          }
-          input:focus {
-            outline: none;
-          }
-          button {
-            border: none;
-            outline: none;
-            background: none;
-            cursor: pointer;
-          }
-          a {
-            text-decoration: none;
-            color: #000;
-          }
-          @font-face {
-            font-family: 'PyeongChangPeace-Bold';
-            src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2206-02@1.0/PyeongChangPeace-Bold.woff2')
-              format('woff2');
-            font-weight: normal;
-            font-style: normal;
-          }
-          body {
-            font-family: 'Roboto', sans-serif;
-          }
-          * {
-            -webkit-tap-highlight-color: transparent !important;
-          }
-        `}
-      />
-    </>
+    <Global
+      styles={css`
+        ${reset}
+        * {
+          box-sizing: border-box;
+        }
+        input {
+          outline: none;
+          border: none;
+        }
+        input:focus {
+          outline: none;
+        }
+        button {
+          border: none;
+          outline: none;
+          background: none;
+          cursor: pointer;
+        }
+        a {
+          text-decoration: none;
+          color: #000;
+        }
+        @font-face {
+          font-family: 'PyeongChangPeace-Bold';
+          src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2206-02@1.0/PyeongChangPeace-Bold.woff2')
+            format('woff2');
+          font-weight: normal;
+          font-style: normal;
+        }
+        body {
+          font-family: 'Roboto', sans-serif;
+        }
+        * {
+          -webkit-tap-highlight-color: transparent !important;
+        }
+      `}
+    />
   );
 };
