+++ packages/client/src/components/DynamicIsland/DynamicIsland.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import { AnimatePresence, motion, Variants } from 'framer-motion';
-import { glassmorphism, zIndexes } from '~/styles';
+import { glassmorphism, zIndexes } from '~/lib/styles';
 import { Button } from '../common';
 
 interface Props {
