+++ packages/client/src/lib/api/post.ts
@@ -2,7 +2,6 @@ import { API } from '~/constants';
 import {
   Comment,
   CreatePostParams,
-  Post,
   PostList,
   PostStats,
   PostWithStats,
