+++ packages/client/src/lib/styles/palette.ts
@@ -1,0 +1,17 @@
+export type NormalColorType =
+  | 'primary'
+  | 'success'
+  | 'secondary'
+  | 'warning'
+  | 'error';
+
+export const palette = {
+  white: '#ffffff',
+  gray: '#dfe6e9',
+  black: '#000000',
+  primary: '#0072F5',
+  secondary: '#7828C8',
+  success: '#17C964',
+  warning: '#F5A524',
+  error: '#F31260',
+} as const;
