+++ packages/client/src/lib/styles/shared.ts
@@ -1,0 +1,10 @@
+import { css } from '@emotion/react';
+
+export const glassmorphism = css`
+  background: rgba(255, 255, 255, 0.25);
+  box-shadow: 0 2px 12px 0 rgba(100, 100, 100, 0.3);
+  backdrop-filter: blur(4px);
+  -webkit-backdrop-filter: blur(4px);
+  border-radius: 10px;
+  border: 1px solid rgba(255, 255, 255, 0.18);
+`;
