+++ packages/client/src/lib/styles/zIndexes.ts
@@ -1,0 +1,8 @@
+export const zIndexes = {
+  Header: 10,
+  Footer: 10,
+  Modal: 100,
+  ChatInput: 20,
+  DynamicIsland: 20,
+  popper: 20,
+};
