+++ packages/client/src/pages/Home/Home.tsx
@@ -8,7 +8,7 @@ import PostList from './PostList';
 import AsyncBoundary from '~/components/base/AsyncBoundary';
 import ErrorFallback from '~/components/base/ErrorFallback';
 import { MESSAGE } from '~/constants';
-import { mediaQuery } from '~/styles';
+import { mediaQuery } from '~/lib/styles';
 import PostListSkeleton from './PostListSkeleton';
 
 const Post = () => {
