+++ packages/client/src/pages/Home/PostListSkeleton.tsx
@@ -1,6 +1,6 @@
 // components
 import styled from '@emotion/styled';
-import { mediaQuery } from '~/styles';
+import { mediaQuery } from '~/lib/styles';
 import PostCardSkeleton from './PostCardSkeleton';
 
 const PostListSkeleton = () => {
