+++ packages/client/src/pages/Write/Write.tsx
@@ -11,7 +11,7 @@ const Write = () => {
     return () => {
       reset();
     };
-  }, []);
+  }, [reset]);
 
   return (
     <Container>
