--- packages/client/src/styles/index.ts
@@ -1,5 +1,0 @@
-export * from './theme';
-export * from './GlobalStyle';
-export * from './MyThemeProvider';
-export * from './theme';
-export * from './common';
